LVRUBYM-218: Move JWT strategy and failure app to separate files
